-Migrated all classes to map.cpp and map.h -Remove files migrated to map.cpp and map.h
