+++ CppRisk/MainDriver.cpp
@@ -1,5 +1,4 @@
 
-#include "MapLoader.h"
 #include "map.h"
 #include <iostream>
 #include <vector>
@@ -7,33 +6,10 @@
 using namespace std;
 
 
-void testLoadMaps() {
-    MapLoader loader;
-    vector<string> paths = {
-        "MAP.txt",
-        "Map2.txt",
-        "test.map"
-    };
-
-    for (const auto& p : paths) {
-        string err;
-        Map* m = loader.load(p, &err);
-        if (m) {
-            cout << "[OK] Loaded: " << p << "\n";
-            cout << *m << "\n";
-            cout << "  -> isValid(): " << (m->isValid() ? "SUCCESS" : "FAILURE") << "\n";
-            delete m;
-        } else {
-            cout << "[ERR] " << p << " => " << err << "\n";
-        }
-        cout << "------------------------------------------\n";
-    }
-}
-
 int main(void)
 {
 	// run all driver test functions
-	testLoadMaps();
+	//testLoadMaps();
 	//testPlayers();
 	//testOrdersLists();
 	//testCards();
