+++ CppRisk/map.h
@@ -5,10 +5,93 @@
 #include <vector>
 #include <iostream>
 #include <unordered_set>
+#include <unordered_map>
 
-class Territory;
+class Map;
 class Continent;
+class Territory;
 
+//MAP LOADER-----------------------------------------------------------------------------------------------------------------------
+class MapLoader {
+public:
+    MapLoader();
+    MapLoader(const MapLoader& other);
+    MapLoader& operator=(const MapLoader& other);
+    ~MapLoader();
+
+    Map* load(const std::string& path, std::string* errorOut = nullptr) const;
+
+    friend std::ostream& operator<<(std::ostream& os, const MapLoader&);
+
+private:
+    static std::string trim(const std::string& s);
+    static std::vector<std::string> splitCSV(const std::string& line);
+};
+
+//TERRITORY-----------------------------------------------------------------------------------------------------------------------
+class Territory {
+public:
+    
+    Territory(const std::string& n, int xCord, int yCord);
+    Territory(const Territory& other);
+    Territory& operator=(const Territory& other);
+    ~Territory();
+
+    
+    void addAdjacentTerritory(Territory* territory);
+    const std::vector<Territory*>& getAdjacentTerritories() const;
+
+    
+    std::string getName() const;
+    int getX() const;
+    int getY() const;
+    int getNumberOfArmies() const;
+
+    
+    void setName(const std::string& n);
+    void setX(int xCord);
+    void setY(int yCord);
+    void setNumberOfArmies(int numOfArmies);
+
+    
+    friend std::ostream& operator<<(std::ostream& out, const Territory& territory);
+
+private:
+    
+    std::string* name;
+    int* x;
+    int* y;
+    int* numberOfArmies;
+    std::vector<Territory*> adjacentTerritories;
+};
+
+//CONTINENT-----------------------------------------------------------------------------------------------------------------------
+class Continent {
+public:
+    
+    Continent(const std::string& n, int bonus);
+    Continent(const Continent& other);
+    Continent& operator=(const Continent& other);
+    ~Continent();
+
+    
+    void addTerritory(Territory* territory);
+    const std::vector<Territory*>& getTerritories() const;
+
+    
+    int getBonus() const;
+    std::string getName() const;
+
+    
+    friend std::ostream& operator<<(std::ostream& out, const Continent& continent);
+
+private:
+    std::string* name;
+    int* bonus;
+    std::vector<Territory*> territories;
+};
+
+//MAP-----------------------------------------------------------------------------------------------------------------------
 class Map {
 public:
